+++ credsweeper/ml_model/features/char_set.py
@@ -1,16 +1,10 @@
 """Most rules are described in 'Secrets in Source Code: Reducing False Positives Using Machine Learning'."""
 import contextlib
-from abc import ABC, abstractmethod
-from pathlib import Path
-from typing import List, Any, Dict, Tuple, Set
+from typing import Dict
 
-import numpy as np
-
-from credsweeper.common.constants import Base, Chars, CHUNK_SIZE
+from credsweeper.common.constants import Base, Chars
 from credsweeper.credentials import Candidate
-from credsweeper.ml_model.features import Feature
-from credsweeper.utils import Util
-
+from credsweeper.ml_model.features.feature import Feature
 
 
 class CharSet(Feature):
