+++ credsweeper/ml_model/features/is_secret_numeric.py
@@ -1,0 +1,16 @@
+"""Most rules are described in 'Secrets in Source Code: Reducing False Positives Using Machine Learning'."""
+
+from credsweeper.credentials import Candidate
+from credsweeper.ml_model.features.feature import Feature
+
+
+class IsSecretNumeric(Feature):
+    """Feature is true if candidate value is a numerical value."""
+
+    def extract(self, candidate: Candidate) -> bool:
+        try:
+            float(candidate.line_data_list[0].value)
+            return True
+        except ValueError:
+            return False
+
